refactor: Extract pipeline orchestration logic   into separate module
